p4i integration Integrate 83232 from 3.1 to main: Get rid of minGW warnings: passing negative value `-1' for argument passing 4 of `LockFile()' argument of negative value `-1' to `long unsigned int' passing negative value `-1' for argument passing 5 of `LockFile()' argument of negative value `-1' to `long  unsigned int' passing negative value `-1' for argument passing 4 of `UnlockFile()' argument of negative value `-1' to `long unsigned int' passing negative value `-1' for argument passing 5 of `UnlockFile()' argument of negative value `-1' to `long unsigned int'
Could the Windows people check LockFile() and UnlockFile()
are used the way they are meant to be used?

[git-p4: depot-paths = "//depot/qt/main/": change = 83233]

+++ src/tools/qsettings.cpp
@@ -329,7 +329,7 @@ static HANDLE openlock( const QString &name, int /*type*/ )
 	fd = CreateFileA( name.local8Bit(), GENERIC_READ, FILE_SHARE_READ|FILE_SHARE_WRITE, NULL, OPEN_ALWAYS, FILE_ATTRIBUTE_NORMAL, NULL );
     } );
 
-    if ( !LockFile( fd, 0, 0, -1, -1 ) ) {
+    if ( !LockFile( fd, 0, 0, (DWORD)-1, (DWORD)-1 ) ) { // ### (DWORD)-1 ???
 #ifdef QT_CHECK_STATE
 	qWarning( "QSettings: openlock failed!" );
 #endif
@@ -337,12 +337,13 @@ static HANDLE openlock( const QString &name, int /*type*/ )
     return fd;
 }
 
+// ### make this function static?
 void closelock( HANDLE fd )
 {
     if ( !fd )
 	return;
 
-    if ( !UnlockFile( fd, 0, 0, -1, -1 ) ) {
+    if ( !UnlockFile( fd, 0, 0, (DWORD)-1, (DWORD)-1 ) ) { // ### (DWORD)-1 ???
 #ifdef QT_CHECK_STATE
 	qWarning( "QSettings: closelock failed!");
 #endif
